Split out gRPC library instrumentation.
