+++ instrumentation-api/src/main/java/io/opentelemetry/instrumentation/api/tracer/RpcServerTracer.java
@@ -9,8 +9,17 @@ import static io.opentelemetry.OpenTelemetry.getPropagators;
 
 import io.grpc.Context;
 import io.opentelemetry.context.propagation.TextMapPropagator;
+import io.opentelemetry.trace.Tracer;
 
 public abstract class RpcServerTracer<REQUEST> extends BaseTracer {
+
+  protected RpcServerTracer() {
+
+  }
+
+  protected RpcServerTracer(Tracer tracer) {
+    super(tracer);
+  }
 
   protected abstract TextMapPropagator.Getter<REQUEST> getGetter();
