+++ instrumentation/grpc-1.5/auto/grpc-1.5-auto.gradle
@@ -1,0 +1,39 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+apply from: "$rootDir/gradle/instrumentation.gradle"
+
+muzzle {
+  pass {
+    group = "io.grpc"
+    module = "grpc-core"
+    versions = "[1.5.0,)"
+  }
+}
+
+def grpcVersion = '1.5.0'
+
+dependencies {
+  implementation project(':instrumentation:grpc-1.5:library')
+
+  library group: 'io.grpc', name: 'grpc-core', version: grpcVersion
+
+  testLibrary group: 'io.grpc', name: 'grpc-netty', version: grpcVersion
+  testLibrary group: 'io.grpc', name: 'grpc-protobuf', version: grpcVersion
+  testLibrary group: 'io.grpc', name: 'grpc-stub', version: grpcVersion
+
+  testImplementation project(':instrumentation:grpc-1.5:testing')
+}
