+++ instrumentation/grpc-1.5/auto/src/main/java/io/opentelemetry/instrumentation/auto/grpc/v1_5/AbstractGrpcInstrumentation.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.instrumentation.auto.grpc.v1_5;
+
+import io.opentelemetry.javaagent.tooling.Instrumenter;
+
+abstract class AbstractGrpcInstrumentation extends Instrumenter.Default {
+
+  public AbstractGrpcInstrumentation() {
+    super("grpc");
+  }
+
+  @Override
+  public final String[] helperClassNames() {
+    return new String[] {
+      "io.opentelemetry.instrumentation.grpc.v1_5.common.GrpcHelper",
+      "io.opentelemetry.instrumentation.grpc.v1_5.client.GrpcClientTracer",
+      "io.opentelemetry.instrumentation.grpc.v1_5.client.GrpcInjectAdapter",
+      "io.opentelemetry.instrumentation.grpc.v1_5.client.TracingClientInterceptor",
+      "io.opentelemetry.instrumentation.grpc.v1_5.client.TracingClientInterceptor$TracingClientCall",
+      "io.opentelemetry.instrumentation.grpc.v1_5.client.TracingClientInterceptor$TracingClientCallListener",
+      "io.opentelemetry.instrumentation.grpc.v1_5.server.GrpcExtractAdapter",
+      "io.opentelemetry.instrumentation.grpc.v1_5.server.GrpcServerTracer",
+      "io.opentelemetry.instrumentation.grpc.v1_5.server.TracingServerInterceptor",
+      "io.opentelemetry.instrumentation.grpc.v1_5.server.TracingServerInterceptor$TracingServerCall",
+      "io.opentelemetry.instrumentation.grpc.v1_5.server.TracingServerInterceptor$TracingServerCallListener",
+    };
+  }
+}
