+++ instrumentation/grpc-1.5/auto/src/main/java/io/opentelemetry/instrumentation/auto/grpc/v1_5/GrpcClientBuilderBuildInstrumentation.java
@@ -1,0 +1,58 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.instrumentation.auto.grpc.v1_5;
+
+import static io.opentelemetry.javaagent.tooling.bytebuddy.matcher.AgentElementMatchers.extendsClass;
+import static java.util.Collections.singletonMap;
+import static net.bytebuddy.matcher.ElementMatchers.declaresField;
+import static net.bytebuddy.matcher.ElementMatchers.isMethod;
+import static net.bytebuddy.matcher.ElementMatchers.named;
+
+import com.google.auto.service.AutoService;
+import io.grpc.ClientInterceptor;
+import io.opentelemetry.instrumentation.grpc.v1_5.client.TracingClientInterceptor;
+import io.opentelemetry.javaagent.tooling.Instrumenter;
+import java.util.List;
+import java.util.Map;
+import net.bytebuddy.asm.Advice;
+import net.bytebuddy.description.method.MethodDescription;
+import net.bytebuddy.description.type.TypeDescription;
+import net.bytebuddy.matcher.ElementMatcher;
+
+@AutoService(Instrumenter.class)
+public class GrpcClientBuilderBuildInstrumentation extends AbstractGrpcInstrumentation {
+
+  @Override
+  public ElementMatcher<TypeDescription> typeMatcher() {
+    return extendsClass(named("io.grpc.ManagedChannelBuilder"))
+        .and(declaresField(named("interceptors")));
+  }
+
+  @Override
+  public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
+    return singletonMap(
+        isMethod().and(named("build")),
+        GrpcClientBuilderBuildInstrumentation.class.getName() + "$AddInterceptorAdvice");
+  }
+
+  public static class AddInterceptorAdvice {
+
+    @Advice.OnMethodEnter(suppress = Throwable.class)
+    public static void addInterceptor(
+        @Advice.FieldValue("interceptors") List<ClientInterceptor> interceptors) {
+      boolean shouldRegister = true;
+      for (ClientInterceptor interceptor : interceptors) {
+        if (interceptor instanceof TracingClientInterceptor) {
+          shouldRegister = false;
+          break;
+        }
+      }
+      if (shouldRegister) {
+        interceptors.add(0, TracingClientInterceptor.newInterceptor());
+      }
+    }
+  }
+}
