--- instrumentation/grpc-1.5/src/main/java/io/opentelemetry/instrumentation/auto/grpc/client/GrpcClientBuilderBuildInstrumentation.java
@@ -1,65 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.instrumentation.auto.grpc.client;
-
-import static io.opentelemetry.javaagent.tooling.bytebuddy.matcher.AgentElementMatchers.extendsClass;
-import static java.util.Collections.singletonMap;
-import static net.bytebuddy.matcher.ElementMatchers.declaresField;
-import static net.bytebuddy.matcher.ElementMatchers.isMethod;
-import static net.bytebuddy.matcher.ElementMatchers.named;
-
-import com.google.auto.service.AutoService;
-import io.grpc.ClientInterceptor;
-import io.grpc.ManagedChannelBuilder;
-import io.opentelemetry.instrumentation.auto.api.ContextStore;
-import io.opentelemetry.instrumentation.auto.api.InstrumentationContext;
-import io.opentelemetry.javaagent.tooling.Instrumenter;
-import java.net.InetSocketAddress;
-import java.util.List;
-import java.util.Map;
-import net.bytebuddy.asm.Advice;
-import net.bytebuddy.description.method.MethodDescription;
-import net.bytebuddy.description.type.TypeDescription;
-import net.bytebuddy.matcher.ElementMatcher;
-
-@AutoService(Instrumenter.class)
-public class GrpcClientBuilderBuildInstrumentation extends AbstractGrpcClientInstrumentation {
-
-  @Override
-  public ElementMatcher<TypeDescription> typeMatcher() {
-    return extendsClass(named("io.grpc.ManagedChannelBuilder"))
-        .and(declaresField(named("interceptors")));
-  }
-
-  @Override
-  public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
-    return singletonMap(
-        isMethod().and(named("build")),
-        GrpcClientBuilderBuildInstrumentation.class.getName() + "$AddInterceptorAdvice");
-  }
-
-  public static class AddInterceptorAdvice {
-
-    @Advice.OnMethodEnter(suppress = Throwable.class)
-    public static void addInterceptor(
-        @Advice.This ManagedChannelBuilder thiz,
-        @Advice.FieldValue("interceptors") List<ClientInterceptor> interceptors) {
-      boolean shouldRegister = true;
-      for (ClientInterceptor interceptor : interceptors) {
-        if (interceptor instanceof TracingClientInterceptor) {
-          shouldRegister = false;
-          break;
-        }
-      }
-      if (shouldRegister) {
-        ContextStore<ManagedChannelBuilder, InetSocketAddress> contextStore =
-            InstrumentationContext.get(ManagedChannelBuilder.class, InetSocketAddress.class);
-        InetSocketAddress sockAddr = contextStore.get(thiz);
-        interceptors.add(0, new TracingClientInterceptor(sockAddr));
-      }
-    }
-  }
-}
