+++ instrumentation/grpc-1.5/testing/src/main/proto/helloworld.proto
