+++ settings.gradle
@@ -88,7 +88,9 @@ include ':instrumentation:geode-1.4'
 include ':instrumentation:google-http-client-1.19'
 include ':instrumentation:grizzly-2.0'
 include ':instrumentation:grizzly-client-1.9'
-include ':instrumentation:grpc-1.5'
+include ':instrumentation:grpc-1.5:auto'
+include ':instrumentation:grpc-1.5:library'
+include ':instrumentation:grpc-1.5:testing'
 include ':instrumentation:guava-10.0'
 include ':instrumentation:hibernate:hibernate-3.3'
 include ':instrumentation:hibernate:hibernate-4.0'
